(LedgerStore) Create ColumnMetrics trait for CF metric reporting
